feat: warn user about Wayland environment
Nobody has been trapped yet, but we're just doing it in case it happens.
Wayland is always the tricky one; they won't let apps know where their
windows are, nor do they let them position themselves. ProbeScope uses
this ability to position its plot controls and plot area selectors,
possibly more in the future. This will work on Windows, macOS, X11 and
simply not on Wayland. For the sake of being simple just ditch Wayland
support as Linux desktop itself is all of a joke (X11 is not, it's the
result of the gold era of Unices)